requirements.txt
# Audio processing
# Removing audioop-lts as it requires Python 3.13+
# Using built-in audioop instead
PyAudio==0.2.14
pvporcupine==3.0.3
SpeechRecognition==3.12.0
playsound==1.2.2

# AI and API integrations
openai>=1.0.0
pydantic>=2.10.5
pydantic-ai>=0.0.19
numpy>=1.24.0

# Google API related
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6

# Utilities
python-dotenv>=1.0.1
typing_extensions>=4.12.2
